Enhanced diagnostics using child diagnostics in sign gutter
* RTags diagnostics can have `children`, which contains extra info such
as the code path that an error came from.  E.g. incorrect constructor
arguments to `make_unique` will show up in the diagnostics for
`unique_ptr`, rather than where `make_unique` was called from.
* So iterate over all diagnostics, rather than just for a given file,
looking for `children` that match the current buffer, and place a sign
in the buffer to indicate it's contribution to the error.